changed price to float, updated stumptown to have error checking, cleaned up

+++ models.py
@@ -6,7 +6,7 @@ class Coffee(ndb.Model):
     description = ndb.StringProperty()
     date_added = ndb.DateTimeProperty(auto_now_add=True)
     date_removed = ndb.DateTimeProperty()
-    price = ndb.IntegerProperty()
+    price = ndb.FloatProperty()
     notes = ndb.StringProperty()
     region = ndb.StringProperty()
     status = ndb.StringProperty()
